move utils to files with clearer names
