+++ packages/nodes-base/nodes/RespondToWebhook/utils/binary.ts
@@ -1,45 +1,8 @@
 import type { IBinaryData, IDataObject, IN8nHttpResponse } from 'n8n-workflow';
 import { BINARY_ENCODING } from 'n8n-workflow';
-import type { Readable } from 'node:stream';
+import type { Readable } from 'stream';
 
-export const configuredOutputs = (
-	version: number,
-	parameters: { enableResponseOutput?: boolean },
-) => {
-	const multipleOutputs = version === 1.3 || (version >= 1.4 && parameters.enableResponseOutput);
-	if (multipleOutputs) {
-		return [
-			{
-				type: 'main',
-				displayName: 'Input Data',
-			},
-			{
-				type: 'main',
-				displayName: 'Response',
-			},
-		];
-	}
-
-	return ['main'];
-};
-
-export const replaceSingleQuotes = (responseBody: string) => {
-	return responseBody.replace(/'/g, '&#39;');
-};
-
-/*
- * To prevent XSS and CSRF, we sanitize the response body by wrapping it in an iframe.
- * This prevents the API from being accessed since the iframe origin will be empty.
- */
-export const sanitizeResponseData = (responseBody: string) => {
-	const parsedResponseBody = replaceSingleQuotes(responseBody);
-
-	return `
-		<iframe srcdoc='${parsedResponseBody}'
-			style="position:fixed; top:0; left:0; width:100vw; height:100vh; border:none; overflow:auto;"
-			allowtransparency="true">
-		</iframe>`;
-};
+import { sanitizeResponseData } from './sanitization';
 
 export const getBinaryResponse = (binaryData: IBinaryData, headers: IDataObject) => {
 	const mimeTypesToSanitize = ['text/html', 'application/xhtml+xml'];
