+++ packages/nodes-base/nodes/RespondToWebhook/utils/outputs.ts
@@ -1,0 +1,20 @@
+export const configuredOutputs = (
+	version: number,
+	parameters: { enableResponseOutput?: boolean },
+) => {
+	const multipleOutputs = version === 1.3 || (version >= 1.4 && parameters.enableResponseOutput);
+	if (multipleOutputs) {
+		return [
+			{
+				type: 'main',
+				displayName: 'Input Data',
+			},
+			{
+				type: 'main',
+				displayName: 'Response',
+			},
+		];
+	}
+
+	return ['main'];
+};
